litelist v0.0.1 Litelist.Moderation

The Moderation context.

Functions

change_flag(flag)

Returns an %Ecto.Changeset{} for tracking flag changes.

Examples

iex> change_flag(flag)
%Ecto.Changeset{source: %Flag{}}

create_flag(attrs \\ %{})

Creates a flag.

Examples

iex> create_flag(%{field: value})
{:ok, %Flag{}}

iex> create_flag(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_flag(flag)

Deletes a Flag.

Examples

iex> delete_flag(flag)
{:ok, %Flag{}}

iex> delete_flag(flag)
{:error, %Ecto.Changeset{}}

get_flag!(id)

Gets a single flag. Post is preloaded

Raises Ecto.NoResultsError if the Flag does not exist.

Examples

iex> get_flag!(123)
%Flag{}

iex> get_flag!(456)
** (Ecto.NoResultsError)

list_flags()

Returns the list of flags. Posts are preloaded

Examples

iex> list_flags()
[%Flag{}, ...]

list_flags_by_neighbor(neighbor)

Returns the list of flags created by the neighbor Posts are preloaded

Examples

iex> list_flags_by_neighbor(neighbor)
[%Flag{}, ...]

list_my_flagged_posts(neighbor)

Returns the list of flags for posts created by the neighbor Posts are preloaded

Examples

iex> list_my_flagged_posts(neighbor)
[%Flag{}, ...]

update_flag(flag, attrs)

Updates a flag.

Examples

iex> update_flag(flag, %{field: new_value})
{:ok, %Flag{}}

iex> update_flag(flag, %{field: bad_value})
{:error, %Ecto.Changeset{}}