litelist v0.0.1 Litelist.Settings.SettingsDatabase

Functions

__using__(opts) (macro)

Alias all the table names in the current scope and require what’s needed.

create(copying \\ [])

create(Amnesia.Table.c()) :: [Amnesia.Table.o()]

Create the database, it calls .create on every defined table.

create!(copying \\ [])

create!(Amnesia.Table.c()) :: [Amnesia.Table.o()]

Create the database, it calls .create on every defined table, raises in case of error.

destroy()

destroy() :: [Amnesia.Table.o()]

Destroy the database, it calls .destroy on every defined table.

destroy!()

destroy!() :: [Amnesia.Table.o()]

Destroy the database, it calls .destroy on every defined table, raises in case of error.

get_settings()

Gets the current settings. Creates default settings if current settings don’t exist. Returns %Settings{}

Examples

iex> get_settings() %Settings{…}

metadata()

metadata() :: Amnesia.Metadata.t()

tables()

tables() :: [atom()]

List of atoms of the defined tables in the database.

update_settings(attrs)

Updates settings.

Examples

iex> update_settings(attrs) %Settings{attrs}

wait(timeout \\ :infinity)

wait(integer() | :infinity) :: :ok | {:timeout, [atom()]} | {:error, atom()}

Wait for the database to be loaded.