litelist v0.0.1 LitelistWeb.FormHelpers

Form Helpers are how LiteList can use a single form for all of its resources.

Functions

get_form_data(type, resource \\ nil)

Determines what resource is being built.

Parameters

type: The Type of Post being created or edited.

resource: Optional. If its an edit form, pass through the object from the controller to the view to this helper.

Examples

iex> LitelistWeb.FormHelpers.get_form_data("ForSale",@current_post)