litelist v0.0.1 LitelistWeb.Utils.SharedUtils

ForSale Utility functions

Functions

add_generated_params(params, atom)

add_generated_params() adds params like current_neighbor for update

add_generated_params(params, conn, type, atom)

add_generated_params() adds params like current_neighbor for create

admin?(neighbor)

admin?(neighbor) If neighbor.admin, return true. Otherwise return false

match_type?(resource, type)

match_type?(resource, type) tests if a resource was created by the neighbor, and if the given type matches the resource

permission?(neighbor, resource, type)

permission?(neighbor, resource, type) tests if a resource was created by the neighbor, and if the given type matches the resource

permitted_params(params, whitelist)

permitted_params(params, whitelist) Removes any non whitelisted attrs from params

resource_owner?(neighbor, resource)

resource_owner?(neighbor, resource) Returns true if neighbor.id == resource.neighbor_id

resource_owner_and_match_type?(neighbor, resource, type)

resource_owner_and_match_type?(neighbor, resource, type) ensures that the resource is owned by the neighbor, and that the types match (job != for_sale)

slugify(string)

slugify iex> slugify(nil) nil iex> slugify(“Some String”) “some-string”