litelist v0.0.1 LitelistWeb.Utils.SharedUtils

ForSale Utility functions

Functions

add_default_status(params, default_status)

adds a status attribute to a map with a given (default) value

add_generated_params(params, atom)

add_generated_params() adds params like current_neighbor for update

add_generated_params(params, conn, type, atom)

add_generated_params() adds params like current_neighbor for create

add_neighbor_id(params, conn)

add_neighbor_id(params, con) Adds the current_neighbor id to params

add_neighbor_id_if_exists(params, conn)

add_neighbor_id_if_exists(params, con) Adds the current_neighbor id to params

add_slug(params)

adds a slug attribute to a passed in map

admin?(neighbor)

admin?(neighbor) If neighbor.admin, return true. Otherwise return false

match_type?(resource, type)

match_type?(resource, type) tests if a resource was created by the neighbor, and if the given type matches the resource

parse_multi_select(params, field)

Parses a multiselect so the field becomes a string (from a List) parse_multi_select(params, field)

permission?(neighbor, resource, type)

permission?(neighbor, resource, type) tests if a resource was created by the neighbor, and if the given type matches the resource

permitted_params(params, whitelist)

permitted_params(params, whitelist) Removes any non whitelisted attrs from params

resource_owner?(neighbor, resource)

resource_owner?(neighbor, resource) Returns true if neighbor.id == resource.neighbor_id

resource_owner_and_match_type?(neighbor, resource, type)

resource_owner_and_match_type?(neighbor, resource, type) ensures that the resource is owned by the neighbor, and that the types match (job != for_sale)

slugify(string)

slugify iex> slugify(nil) nil iex> slugify(“Some String”) “some-string”

update_slug(params)

updates a slug attribute to a passed in map